uncore: Add enable for CtrlXing and expose it to the top level
CtrlXing = ControlledCrossing

According to the generated verilog code, although CtrlXing is parallel
to the `val module' of RocketTile, it is indeed a submodule of RocketTile.
Therefore, if we want to upward the enable port to wire at PARTFPGATop level,
we should at least provide enable in RocketTiles io ports in order to access
the inner CtrlXing enable port.

To avoid Chisel3 error on sink unavailable in RocketTile module,
the reference to the instance of ControlledCrossing is obtained
by RocketTile with a concrete name to refer to its module.io

According to a comment
(https://github.com/ucb-bar/chisel3/issues/350#issuecomment-258077010),
`import Chisel._' will pass errors to Firttl, which is inconvenient to
figure out the problem. Therefore the added module is changed to use
`import chisel3._'