Fixed the L2 "accepted release for block not resident in cache".
This was how it happened:
Terminology:
Cached acquire: acquire to L2 Cache
Uncached acquire: acquire to uncacheable memory, usually mmio.
Currently, L1 DCache:
1. disallow multiple cached acquire inflight
2. allow multiple uncached acquire inflight
3. cached and uncached acquire can go seperately

Here is the corner case, that caused spurious release on L2.
1. acquire cached block 1(addr: 0x80003dc0), miss
2. grant and update meta data for block 1(way_en: 2 idx: 37 isValid: 1 tag: 80003)
3. acquire uncached block 2(addr: 0x60000004 uart mmio)
4. acquire cache block 3(addr: 0x80003e00), miss, it stores way_en and paddr on pipeline stage 2
5. access cache block 1(addr: 0x80003df8), since one cached acquire is
inflight, it is stalled on pipeline stage 1
6. grant uart mmio, which caused paddr on s2 stag updated by paddr in
stage 1, which is 0x80003df8(!!!!!!!!!)
7. grant acquire 3 and update meta data 3(way_en: 8, addr: 0x80003df8, tag: 80003),
which means in set 0x37, there are two ways with the same tag!!!!!
8. then, the cache block will be released two times, which caused
assertion fail in L2 voluntary release tracker.

Fix: we temporarily stall s1 stage when there is an
acquires(cached/uncached) inflight.